+++ ResortManagementApp/Models/Administration/DTOS/ExtraChargersDto.cs
@@ -9,10 +9,9 @@ namespace ResortManagementApp.Models.Administration.DTOS
         public Guid Id { get; set; }
        
         public string ExtraChargeType { get; set; }
-        public double? Amount { get; set; }
-
-        public double? RateUSD { get; set; }
-        public double? RateLKR { get; set; }
+        public decimal? Amount { get; set; }
+        public decimal? RateUSD { get; set; }
+        public decimal? RateLKR { get; set; }
         public string?  Comment { get; set; }
         public string? Details { get; set; }
         public DateTime? DateTime { get; set; }
